pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.31.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "nltk>=3.9.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
]

[tool.uv.sources]
src = { path = "src" }
